test: keep sharp out of integration files that only read the screen, and pin the missing-disc error
